Stuck
Having issues with my Nav bar not working. When clicked the address bar shows the appropriate routes but nothing loads on the page. Also, I created the NotFound component, imported it into my App.js and added it as a Route in the render method but it does not work as well. Help needed!